pyproject.toml
[build-system]
requires = ["setuptools >= 66.0.0"]
build-backend = "setuptools.build_meta"

[project]
version = "3.0.0"
name = "krml"
description = "Knowledge Representation Markup Language"
keywords = [
    "Knowledge graphs",
    "Markdown",
    "RDF"
]
readme = 'README.md'
urls = { project-url = "https://github.com/edwardanderson/krml" }
dependencies = [
    "lxml",
    "markdown-it-py[plugins]",
    "python-frontmatter",
    "python-testmark",
    "rdflib",
    "rich",
    "saxonche",
    "typer"
]

[tool.setuptools.packages.find]

[project.scripts]
krml = "krml.cli:app"
